Minfactor from int to double

+++ src/Config.h
@@ -31,7 +31,7 @@ public:
 	int SimpleLengthFactor;
 	int numStab;
 	int Hifactor;
-	int Minfactor;
+	double Minfactor;
 	double Hcrit_global;
 	int transport_eq;
 	int alpha_varies;
